teste da mensagem de audio no server
