+++ API/bot/mensagem.js
@@ -1,9 +1,13 @@
 import text from './text/mensagemTexto.js';
+import audio from './audio/audioMessage.js';
 
 // Objeto para redirecionar as mensagens baseado no tipo da mensagem recebida
 const redirecionar = {
     text: async (consulta, objetoMensagem, res) => {
         return await text(consulta, objetoMensagem.from, objetoMensagem.text.body, res);
+    },
+    audio: async (consulta, objetoMensagem, res) => {
+        return await audio(objetoMensagem.audio.id);
     }
 }
